requirements.txt

# === FastAPI e Servidor ===
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20

# === LangChain e LangGraph ===
langgraph==0.6.8
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-anthropic==0.3.21
langchain-google-genai>=1.0.0
langchain-deepseek==0.1.3

# === Provedores de LLM ===
google-generativeai>=0.5.0
litellm==1.77.5

# === Parsing e Processamento ===
beautifulsoup4==4.14.2
lxml==6.0.2
pyyaml==6.0.2

# === Validação e Settings ===
pydantic==2.10.6
pydantic-settings==2.6.1

# === Logging ===
loguru==0.7.3

# === Utilities ===
python-dateutil==2.9.0
aiofiles==24.1.0

unidecode==1.3.8
